requirements.txt
crayons
netmiko
pyyaml
requests